python/fraud-base-mcp-server/pyproject.toml
[project]
name = "fraud-base-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.7.1",
    "mysql-connector-python>=9.3.0",
]
